update(controller): update batch endpoint

+++ src/main/java/com/equipo_38/flight_on_time/dto/BatchPredictionDTO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record BatchPredictionDTO(
                                  int total,
-                                 int success,
-                                 int failed,
-                                 List<BatchItemDTO> items // <--- CAMBIO CLAVE: Usamos BatchItemDTO en vez de PredictionResponseDTO
+                                 Long success,
+                                 Long failed,
+                                 List<BatchItemDTO> items
 ) {}
